Fixed a bug causing ruber to look for plugins in the installation directory of a previous version
If the user changes the plugin directories settings, maybe adding one
directory, with version x.y.z of ruber installed, the setting is written
to the config file, with an explicit reference to the installation
directory of version x.y.z.

When another version a.b.c of ruber is installed, the default value for
the plugin directories wouldn't be used anymore (because the setting
already has a value), so Ruber will still look for plugins in the
installation directory of version x.y.z.

What is done now is to have Ruber::Application#load_settings replace
lines looking like previous installation directories with the
installation directory for the current version.

Also, now you don't access the general/plugin_dirs setting directly, but
only using Application#plugin_dirs and Application#plugin_dirs=