Made ProjectFilesList#file_in_project? URL aware
It now returns false for URLs pointing to remote files.

Note that this isn't done by creating a KDE::Url object, but rather by
checking whether the name of the file looks like an URL, that is whether
it begins with any number of letters and numbers followed by a colon and
two slashes. The URL specification requires the slashes only if the URL
as an authority part, but, as far as I understand, files will always
have it, so this shouldn't be a problem