Added support for documents without view
Before, it was (mostly) assumed that each document had a view. Now, this
is not warranteed anymore. Anyone wanting to use the view associated
with a document must always check that document.view is not nil, or
create one at need using MainWindow#editor_for!.

Also, the state plugin has been changed so that it records separately
which documents were open and which also had a view. All open documents
will be re-opened, but a view will only be created for those which
already had it