Removed the RubyRunner::RubyRunnerPluginInternal hack
Now that load can be used instead of require for files listed in the
require section of the PSF, there's no need to use this hack. The
RubyRunner::RubyRunnerPlugin class is now the true class, its name is
removed when the plugin is unloaded and, because the file is listed with
the extension in the PSF, it'll be loaded again if the plugin is
reloaded.