Added pin button to output widget and changed behaviour on Meta key
Now to open a file without closing the tool widget, you need to pin it
down using the button.

Clicking on a file name with the meta key pressed a new editor is
created even if another one already exists. The hints used are those
returned by the hints method.

When clicking on a file name without pressing the Meta key, an existing
editor is used, if possible. The general/tool_open_file option is only
used when creating a new editor