Made horizontal scrollbar work in the RSpec plugin
Apparently, horizontal scrollbars aren't automatically shown if a
column's contents are larger than the column's size, but only if the
column size is larger than the availlable space. So, when an item is
expanded, we resize the second column (the one which displays the
backtrace and the only one which is likely to require scrollbar). The
same is done when an item is collapsed, so that the scrollbar is hidden