Attempt to fix problem with auto_end when inserting newline after existing keyword
With the old behaviour, an end was always inserted. Now, instead, it
checks whether the number of leading whitespaces of the line where the
newline was inserted is lower or greater than the number of leading
whitespaces in the following line. If it is lower, it means that the
following line is more indented than the previous, so we assume it is
already part of an existing structure (module, class, def, if...) and do
not insert an end