Made ComponentManager iterate on components in loading order by default
ComponentManager#each_component and ComponentManager#each_plugin used to
iterate on components in reverse loading order (last loaded plugins were
passed first). This can be unexpected, so I made the loading order the
default iteration order. To have the old behaviour (which is useful in
some situations) you have to pass a :reverse argument to the methods