Workaround for a crash when closing session
After the introduction of World component, there is a crash at session
closing. The crash is on the C++ side, in
KateViewInternal::~KateViewInternal and doesn't happen when quitting
Ruber either using File/Quit or clicking on the X button on the title
bar.

I can't fully understand the cause of the bug, and I suspect that to
address it correctly, I'd need to change how application closing works.
The workaround I've applied for the time being is to explicitly hide the
active editor from a method called in case of a closing session. Why
this should have any influence, I don't know.