Correct workaround for bug when closing session
With the new world component, there's a bug at session shutdown which
causes a crash from the C++ side of Ruber (in particular, from
KateViewInternal destructor).

Since I haven't had the time to find out the cause, I tried a
workaround, which is to close all the documents from
MainWindow#queryClose, after calling Application#quit_ruber. Since that
method seems to be called twice at session shutdown, an instance
variable is used to ensure this only happens the second time the method
is called