Move the registration with plugins outside AbstractProject#initialize
Now extensions and options are added in a method called #finalize, which
is not called by AbstractProject#initialize. Project#initialize does
call it, while DocumentProject#initialize does not. It is called by
Document#initialize after creating the project itself.

In the case of DocumentProject, this allows extensions to freely access
the project itself from their #initialize method