Another attempt at having the horizontal scrollbar correctly displayed
Qt::TreeView#resize_column_to_contents doesn't work for spanning
columns, which means it won't take into account the width of spec
descriptions when resizing the columns. In turn, this implies that long
spec descriptions will be cut off at the end.

The workaround I'm using is this:
* when rspec stops running, compute the space needed to draw the text of
	the longest toplevel item (they all are in spanning columns) using the
	view's font meterics
* whenever an item is expanded or collapsed, resize the columns then, if
	the second column is less wide than the size computed above, resize
	it.

The above procedure is also carried on by #load_settings, because if the
font used to draw the text changes, the widths will also change.

Currently, the width needed for toplevel items is computed taking into
account also items which have been filtered out.

PROBLEM: the width of the column should be larger than the space needed
for the text (because there are margins and so on). Unfortunately, I
couldn't find a way to determine this basing on the current style, so I
just added 30 pixels to the column width. It works correctly on my
system with the Oxygen style, but I'm not confident it's a good
solution.