Added dir patterns for specs
Now the pattern to transform a code file name into the name of the
corresponding spec can include the %d and %dn special sequences, where n
can be any number of digits. The former inserts the directory of the
file, relative to the code directory; the latter inserts the n-th part
of that directory. For example, if the file path relative to the code
directory is abc/def/ghi/xyz.rb, %d would insert abc/def/ghi, while %d2
would insert ghi.