Don't evaluate line choosen from the combobox
The input line was sent to irb in response to the
KDE::ComboBox#activated signal, which is also emitted when the user
chooses an entry from the popup list. This shouldn't happen as the user
may want to change the line. To avoid the issue, now the
KDE::ComboBox#returnPressed(QString) signal is used