Workaround for issue with YAML in 1.9.3
The psych YAML library, which is used by default (I think) in ruby 1.9.3
has problems parsing YAML files with hashes and colons. According to the
ruby bug tracker, this is caused by an ambiguity in the YAML specs,
resulting in a possible bug in psych. Until the issue is fixed, the
correct solution would be to fix all YAML files (including plugin files,
so that they work correctly). For the time being, however, just force
the use of the old Syck library.