[Issue 15] Added project_files_only option
On a per-project base, the user can now choose whether the project
browser tool widget should display all files or only project files. That
will be the behaviour when the project is opened the next time. After
that, the user will be able to alter the chosen behaviour using the
context menu, as it was before. The context menu doesn't change the
value of the setting stored in the project file.