Added history to RI plugin
This required a bit of refactoring, leading to a better definition of roles: the
Plugin class performs the search and formats the result, but doesn't take care
of displaying the results. This task is performed by the tool widget, which also
manages history (using an helper class). This way, RI searches can also be
performed by other plugins without showing the results in the tool widget.